Добавила интерфейсы в domain.search, models, изменила Resource чтобы содержало state ошибки, потом пригодиться для отображения ошибок во viewModel
