+++ app/src/main/java/ru/practicum/android/diploma/data/dto/Response.kt
@@ -1,12 +1,6 @@
 package ru.practicum.android.diploma.data.dto
 
-enum class ResponseStatus {
-    SUCCESS, // Успешный запрос (200)
-    NOT_FOUND, // Ничего не найдено (404)
-    SERVER_ERROR, // Внутренняя ошибка сервера (500)
-    UNKNOWN_ERROR, // Неизвестная ошибка
-    NO_INTERNET, // Нет интернета
-}
+import ru.practicum.android.diploma.util.ResponseStatus
 
 open class Response {
     var status: ResponseStatus = ResponseStatus.UNKNOWN_ERROR
