+++ app/src/main/java/ru/practicum/android/diploma/domain/search/VacanciesInteractor.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.search
+
+import ru.practicum.android.diploma.domain.search.models.VacanciesSearchResult
+import ru.practicum.android.diploma.util.Resource
+
+interface VacanciesInteractor {
+    suspend fun searchVacancies(expression: String, page: Int): Resource<VacanciesSearchResult>
+}
