+++ app/src/main/java/ru/practicum/android/diploma/domain/search/models/VacanciesSearchResult.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.search.models
+
+data class VacanciesSearchResult(
+    val found: Int,
+    val pages: Int,
+    val page: Int,
+    val items: List<VacancyPreview>
+)
