+++ app/src/main/java/ru/practicum/android/diploma/domain/search/models/VacancyPreview.kt
@@ -1,0 +1,12 @@
+package ru.practicum.android.diploma.domain.search.models
+
+data class VacancyPreview(
+    val id: String,
+    val name: String,
+    val salaryCurrency: String?,
+    val salaryFrom: String?,
+    val salaryTo: String?,
+    val addressCity: String?,
+    val employerName: String
+)
+
